+++ app/Http/Controllers/ApitokenController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Apitoken;
+use App\Models\Permission;
 use Illuminate\Http\Request;
 
 class ApitokenController extends Controller
@@ -10,29 +11,43 @@ class ApitokenController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\Response|\Illuminate\Http\JsonResponse
      */
-    public function index()
+    public function index(Request $request)
     {
+        $token = Apitoken::where('token','=',$request->bearerToken())->first();
+        if(!$token->role->permissions->contains(Permission::firstWhere('name', '=', 'tokens_viewAny')))
+        {
+            return response('Du har ikke de fornødne tilladelser', 403);
+        }
+
         $data = Apitoken::orderBy('id','ASC')->get();
         if(count($data) == 0){
             return response('Ingen API Tokens', 404);
         }
-        return $data;
+
+        return response()->json($data);
     }
 
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Database\Eloquent\Builder[]|\Illuminate\Database\Eloquent\Collection|\Illuminate\Database\Query\Builder[]|\Illuminate\Support\Collection
+     * @return \Illuminate\Http\Response|\Illuminate\Http\JsonResponse
      */
-    public function deleted()
+    public function deleted(Request $request)
     {
+        $token = Apitoken::where('token','=',$request->bearerToken())->first();
+        if(!$token->role->permissions->contains(Permission::firstWhere('name', '=', 'tokens_viewAny_deleted')))
+        {
+            return response('Du har ikke de fornødne tilladelser', 403);
+        }
+
         $data = Apitoken::onlyTrashed()->orderBy('id','ASC')->get();
         if(count($data) == 0){
             return response('Ingen Slettede API Tokens', 404);
         }
-        return $data;
+
+        return response()->json($data);
     }
 
     /**
@@ -43,30 +58,53 @@ class ApitokenController extends Controller
      */
     public function store(Request $request)
     {
+        $token = Apitoken::where('token','=',$request->bearerToken())->first();
+        if(!$token->role->permissions->contains(Permission::firstWhere('name', '=', 'tokens_create')))
+        {
+            return response('Du har ikke de fornødne tilladelser', 403);
+        }
+
         $data = (new Apitoken());
+
         $timestamp = now();
         $data->token = hash('sha512',$timestamp);
-        $data->target_id = $request->target_id;
-        $data->role_id = $request->role_id;
-        $data->expires_at = $request->expires_at;
+
+        if(isset($request->target_id)){
+            $data->target_id = $request->target_id;
+        }
+
+        if(isset($request->role_id)){
+            $data->role_id = $request->role_id;
+        }
+
+        if(isset($request->expires_at)){
+            $data->expire_at = $request->expires_at;
+        }
+
         $data->save();
 
-        return response('API Token oprettet med id: '.$data->id , 202);
+        return response('API Token oprettet med id: '.$data->id , 201);
     }
 
     /**
      * Display the specified resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\Response|\Illuminate\Http\JsonResponse
      */
-    public function show($id)
+    public function show(Request $request, $id)
     {
+        $token = Apitoken::where('token','=',$request->bearerToken())->first();
+        if(!$token->role->permissions->contains(Permission::firstWhere('name', '=', 'tokens_view')))
+        {
+            return response('Du har ikke de fornødne tilladelser', 403);
+        }
+
         $data = Apitoken::where('id','=',$id)->first();
         if(!$data){
             return response('API Token ikke fundet', 404);
         }
 
-        return $data;
+        return response()->json($data);
     }
 
     /**
@@ -77,14 +115,31 @@ class ApitokenController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $token = Apitoken::where('token','=',$request->bearerToken())->first();
+        if(!$token->role->permissions->contains(Permission::firstWhere('name', '=', 'tokens_edit')))
+        {
+            return response('Du har ikke de fornødne tilladelser', 403);
+        }
+
         $data = Apitoken::where('id','=',$id)->first();
         if(!$data){
             return response('API Token ikke fundet', 404);
         }
-        $data->target_id = $request->target_id;
-        $data->role_id = $request->role_id;
-        $data->expire_at = $request->expires_at;
+
+        if(isset($request->target_id)){
+            $data->target_id = $request->target_id;
+        }
+
+        if(isset($request->role_id)){
+            $data->role_id = $request->role_id;
+        }
+
+        if(isset($request->expires_at)){
+            $data->expire_at = $request->expires_at;
+        }
+
         $data->save();
+
         return response('API Token opdateret', 200);
     }
 
@@ -93,13 +148,21 @@ class ApitokenController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
+        $token = Apitoken::where('token','=',$request->bearerToken())->first();
+        if(!$token->role->permissions->contains(Permission::firstWhere('name', '=', 'tokens_delete')))
+        {
+            return response('Du har ikke de fornødne tilladelser', 403);
+        }
+
         $data = Apitoken::where('id','=',$id)->first();
         if(!$data){
             return response('API Token ikke fundet', 404);
         }
+
         $data->delete();
+
         return response('API Token slettet', 204);
     }
 
@@ -108,13 +171,21 @@ class ApitokenController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function delete_force($id)
+    public function delete_force(Request $request, $id)
     {
+        $token = Apitoken::where('token','=',$request->bearerToken())->first();
+        if(!$token->role->permissions->contains(Permission::firstWhere('name', '=', 'tokens_delete_force')))
+        {
+            return response('Du har ikke de fornødne tilladelser', 403);
+        }
+
         $data = Apitoken::onlyTrashed()->where('id','=',$id)->first();
         if(!$data){
             return response('API Token ikke fundet', 404);
         }
+
         $data->forceDelete();
+
         return response('API Token permanent slettet', 204);
     }
 
@@ -123,13 +194,21 @@ class ApitokenController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function restore($id)
+    public function restore(Request $request, $id)
     {
+        $token = Apitoken::where('token','=',$request->bearerToken())->first();
+        if(!$token->role->permissions->contains(Permission::firstWhere('name', '=', 'tokens_restore')))
+        {
+            return response('Du har ikke de fornødne tilladelser', 403);
+        }
+
         $data = Apitoken::withTrashed()->where('id','=',$id)->first();
         if(!$data){
             return response('API Token ikke fundet', 404);
         }
+
         $data->restore();
-        return response('API Token genoprettet', 204);
+
+        return response('API Token genoprettet', 200);
     }
 }
