+++ app/Http/Controllers/ZoneController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Apitarget;
 use App\Models\Apitoken;
+use App\Models\Permission;
 use App\Models\User;
 use App\Models\Zone;
 use Illuminate\Http\Request;
@@ -13,29 +14,43 @@ class ZoneController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\Response|\Illuminate\Http\JsonResponse
      */
-    public function index()
+    public function index(Request $request)
     {
+        $token = Apitoken::where('token','=',$request->bearerToken())->first();
+        if(!$token->role->permissions->contains(Permission::firstWhere('name', '=', 'zones_viewAny')))
+        {
+            return response('Du har ikke de fornødne tilladelser', 403);
+        }
+
         $data = Zone::orderBy('id','ASC')->get();
         if(count($data) == 0){
             return response('Ingen Zoner', 404);
         }
-        return $data;
+
+        return response()->json($data);
     }
 
     /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Database\Eloquent\Builder[]|\Illuminate\Database\Eloquent\Collection|\Illuminate\Database\Query\Builder[]|\Illuminate\Support\Collection
+     * @return \Illuminate\Http\Response|\Illuminate\Http\JsonResponse
      */
-    public function deleted()
+    public function deleted(Request $request)
     {
+        $token = Apitoken::where('token','=',$request->bearerToken())->first();
+        if(!$token->role->permissions->contains(Permission::firstWhere('name', '=', 'zones_viewAny_deleted')))
+        {
+            return response('Du har ikke de fornødne tilladelser', 403);
+        }
+
         $data = Zone::onlyTrashed()->orderBy('id','ASC')->get();
         if(count($data) == 0){
             return response('Ingen Slettede Zoner', 404);
         }
-        return $data;
+
+        return response()->json($data);
     }
 
     /**
@@ -46,26 +61,42 @@ class ZoneController extends Controller
      */
     public function store(Request $request)
     {
+        $token = Apitoken::where('token','=',$request->bearerToken())->first();
+        if(!$token->role->permissions->contains(Permission::firstWhere('name', '=', 'zones_create')))
+        {
+            return response('Du har ikke de fornødne tilladelser', 403);
+        }
+
         $data = (new Zone());
-        $data->name = $request->name;
+
+        if(isset($request->name)){
+            $data->name = $request->name;
+        }
+
         $data->save();
 
-        return response('Zone oprettet med id: '.$data->id , 202);
+        return response('Zone oprettet med id: '.$data->id , 201);
     }
 
     /**
      * Display the specified resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return \Illuminate\Http\Response|\Illuminate\Http\JsonResponse
      */
-    public function show($id)
+    public function show(Request $request, $id)
     {
+        $token = Apitoken::where('token','=',$request->bearerToken())->first();
+        if(!$token->role->permissions->contains(Permission::firstWhere('name', '=', 'zones_view')))
+        {
+            return response('Du har ikke de fornødne tilladelser', 403);
+        }
+
         $data = Zone::where('id','=',$id)->first();
         if(!$data){
             return response('Zone ikke fundet', 404);
         }
 
-        return $data;
+        return response()->json($data);
     }
 
     /**
@@ -76,12 +107,23 @@ class ZoneController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $token = Apitoken::where('token','=',$request->bearerToken())->first();
+        if(!$token->role->permissions->contains(Permission::firstWhere('name', '=', 'zones_edit')))
+        {
+            return response('Du har ikke de fornødne tilladelser', 403);
+        }
+
         $data = Zone::where('id','=',$id)->first();
         if(!$data){
             return response('Zone ikke fundet', 404);
         }
-        $data->name = $request->name;
+
+        if(isset($request->name)){
+            $data->name = $request->name;
+        }
+
         $data->save();
+
         return response('Zone opdateret', 200);
     }
 
@@ -90,13 +132,21 @@ class ZoneController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
+        $token = Apitoken::where('token','=',$request->bearerToken())->first();
+        if(!$token->role->permissions->contains(Permission::firstWhere('name', '=', 'zones_delete')))
+        {
+            return response('Du har ikke de fornødne tilladelser', 403);
+        }
+
         $data = Zone::where('id','=',$id)->first();
         if(!$data){
             return response('Zone ikke fundet', 404);
         }
+
         $data->delete();
+
         return response('Zone slettet', 204);
     }
 
@@ -105,13 +155,21 @@ class ZoneController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function delete_force($id)
+    public function delete_force(Request $request, $id)
     {
+        $token = Apitoken::where('token','=',$request->bearerToken())->first();
+        if(!$token->role->permissions->contains(Permission::firstWhere('name', '=', 'zones_delete_force')))
+        {
+            return response('Du har ikke de fornødne tilladelser', 403);
+        }
+
         $data = Zone::onlyTrashed()->where('id','=',$id)->first();
         if(!$data){
             return response('Zone ikke fundet', 404);
         }
+
         $data->forceDelete();
+
         return response('Zone permanent slettet', 204);
     }
 
@@ -120,19 +178,26 @@ class ZoneController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function restore($id)
+    public function restore(Request $request, $id)
     {
+        $token = Apitoken::where('token','=',$request->bearerToken())->first();
+        if(!$token->role->permissions->contains(Permission::firstWhere('name', '=', 'zones_restore')))
+        {
+            return response('Du har ikke de fornødne tilladelser', 403);
+        }
+
         $data = Zone::withTrashed()->where('id','=',$id)->first();
         if(!$data){
             return response('Zone ikke fundet', 404);
         }
+
         $data->restore();
 
-        return response('Zone genoprettet', 204);
+        return response('Zone genoprettet', 200);
     }
 
     /**
-     * Authenticate the yser.
+     * Authenticate the user.
      *
      * @return \Illuminate\Http\Response
      */
@@ -143,7 +208,7 @@ class ZoneController extends Controller
 
         $user = User::where('email',"=",$email)
             ->where('password',"=",$password)
-            ->first();
+            ->first()
         ;
 
         if(!$user){
